Refactored service initialization and moved to a separate fixture
